+++ src/lib.rs
@@ -1,7 +1,7 @@
-// core.rs
+// lib.rs
 // Copyright (c) 2025 Lunatic Fringers
 // This file is part of "WG-Bridge" under the AGPL-3.0-or-later license.
 // See the LICENSE file in the project root or <https://www.gnu.org/licenses/> for details.
 
-pub mod logger;
-pub mod config;
+
+pub mod core;