+++ CHANGELOG.md
@@ -24,3 +24,4 @@ and this project adheres to
 - Options `connect`, `disconnect`, `list` and `status` to commands
 - 2FA manager that allow to manage it on a connection-by-connection basis
 - Logging errors in the default Linux log directory
+- Configuration parser